update: move item creation to separate hooks
