+++ client/a8c-for-agencies/components/sidebar-menu/purchases.tsx
@@ -1,17 +1,9 @@
 import page from '@automattic/calypso-router';
-import { chevronLeft, key, store, receipt, payment } from '@wordpress/icons';
+import { chevronLeft } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
-import { useMemo } from 'react';
 import Sidebar from '../sidebar';
-import {
-	A4A_OVERVIEW_LINK,
-	A4A_PURCHASES_LINK,
-	A4A_LICENSES_LINK,
-	A4A_BILLING_LINK,
-	A4A_INVOICES_LINK,
-	A4A_PAYMENT_METHODS_LINK,
-} from './lib/constants';
-import { createItem } from './lib/utils';
+import usePurchasesMenuItems from './hooks/use-purchases-menu-items';
+import { A4A_OVERVIEW_LINK, A4A_PURCHASES_LINK } from './lib/constants';
 
 type Props = {
 	path: string;
@@ -19,59 +11,7 @@ type Props = {
 
 export default function ( { path }: Props ) {
 	const translate = useTranslate();
-
-	const menuItems = useMemo( () => {
-		return [
-			createItem(
-				{
-					icon: key,
-					path: A4A_PURCHASES_LINK,
-					link: A4A_LICENSES_LINK,
-					title: translate( 'Licenses' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Purchases / Licenses',
-					},
-				},
-				path
-			),
-			createItem(
-				{
-					icon: store,
-					path: A4A_PURCHASES_LINK,
-					link: A4A_BILLING_LINK,
-					title: translate( 'Billing' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Purchases / Billing',
-					},
-				},
-				path
-			),
-			createItem(
-				{
-					icon: payment,
-					path: A4A_PURCHASES_LINK,
-					link: A4A_PAYMENT_METHODS_LINK,
-					title: translate( 'Payment Methods' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Purchases / Payment Methods',
-					},
-				},
-				path
-			),
-			createItem(
-				{
-					icon: receipt,
-					path: A4A_PURCHASES_LINK,
-					link: A4A_INVOICES_LINK,
-					title: translate( 'Invoices' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Purchases / Invoices',
-					},
-				},
-				path
-			),
-		].map( ( item ) => createItem( item, path ) );
-	}, [ path, translate ] );
+	const menuItems = usePurchasesMenuItems( path );
 
 	return (
 		<Sidebar
