+++ client/a8c-for-agencies/components/sidebar-menu/sites.tsx
@@ -1,16 +1,9 @@
 import page from '@automattic/calypso-router';
-import { category, chevronLeft, starEmpty, warning } from '@wordpress/icons';
+import { chevronLeft } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
-import { useMemo } from 'react';
-import useNoActiveSite from 'calypso/a8c-for-agencies/hooks/use-no-active-site';
 import Sidebar from '../sidebar';
-import {
-	A4A_OVERVIEW_LINK,
-	A4A_SITES_LINK,
-	A4A_SITES_LINK_FAVORITE,
-	A4A_SITES_LINK_NEEDS_ATTENTION,
-} from './lib/constants';
-import { createItem } from './lib/utils';
+import useSitesMenuItems from './hooks/use-sites-menu-items';
+import { A4A_OVERVIEW_LINK, A4A_SITES_LINK } from './lib/constants';
 
 type Props = {
 	path: string;
@@ -18,69 +11,7 @@ type Props = {
 
 export default function ( { path }: Props ) {
 	const translate = useTranslate();
-
-	const noActiveSite = useNoActiveSite();
-
-	const menuItems = useMemo( () => {
-		if ( noActiveSite ) {
-			// We hide the rest of the options when we do not have sites yet.
-			return [
-				createItem(
-					{
-						id: 'sites-all-menu-item',
-						icon: category,
-						path: A4A_SITES_LINK,
-						link: A4A_SITES_LINK,
-						title: translate( 'All' ),
-						trackEventProps: {
-							menu_item: 'Automattic for Agencies / Sites / All',
-						},
-					},
-					path
-				),
-			];
-		}
-
-		return [
-			createItem(
-				{
-					icon: warning,
-					path: A4A_SITES_LINK,
-					link: A4A_SITES_LINK_NEEDS_ATTENTION,
-					title: translate( 'Needs Attention' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Sites / Needs Attention',
-					},
-				},
-				path
-			),
-			createItem(
-				{
-					icon: starEmpty,
-					path: A4A_SITES_LINK,
-					link: A4A_SITES_LINK_FAVORITE,
-					title: translate( 'Favorites' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Sites / Favorites',
-					},
-				},
-				path
-			),
-			createItem(
-				{
-					id: 'sites-all-menu-item',
-					icon: category,
-					path: A4A_SITES_LINK,
-					link: A4A_SITES_LINK,
-					title: translate( 'All' ),
-					trackEventProps: {
-						menu_item: 'Automattic for Agencies / Sites / All',
-					},
-				},
-				path
-			),
-		].map( ( item ) => createItem( item, path ) );
-	}, [ noActiveSite, path, translate ] );
+	const menuItems = useSitesMenuItems( path );
 
 	return (
 		<Sidebar
